backend/requirements.txt
# Django core components
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8

# database
psycopg2-binary>=2.9.0

# authentication and authorization
djangorestframework-simplejwt>=5.2.0

# data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
requests>=2.31.0

# machine learning
tensorflow>=2.12.0
scikit-learn>=1.3.0

# task scheduling
celery>=5.3.0
redis>=4.5.0

# Excel processing
openpyxl>=3.1.0
xlrd>=2.0.0

# development tools
python-dotenv>=1.0.0
whitenoise>=6.5.0 